move arrow stuff to a .c file
